Implement a robot pose setter node that overrides all inputs
Directly manipulate Gazebo model to put the robot where we want to,
by publishing on the /gazebo/set_model_state topic the pose of the
robot.
This may be a solution to problem of using Gazebo controllers, thus
opening a way of using 'perfect controllers' in the simulator.
Still need to try to set link or joint positions this way.

There is still a small ugly hack since we need to publish the new
pose two times or more for it to take effect.